+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [1.1.0] - 2016-03-17
 ### Added
 * SIGTERM will cause OGR and ZIP to exit
 
@@ -91,6 +91,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Initial release: API supports creating VRTs, Geojson, shapefile, kml and csv
 
 [Unreleased]: https://github.com/koopjs/geoxform/compare/v1.0.15...HEAD
+[1.1.0]: https://github.com/koopjs/geoxform/compare/v1.0.15...v1.1.0
 [1.0.15]: https://github.com/koopjs/geoxform/compare/v1.0.14...v1.0.15
 [1.0.14]: https://github.com/koopjs/geoxform/compare/v1.0.13...v1.0.14
 [1.0.13]: https://github.com/koopjs/geoxform/compare/v1.0.12...v1.0.13